requirements.txt
Django==2.2
django-crispy-forms==1.14.0
django-phonenumber-field==6.1.0
mysqlclient==2.1.0
phonenumbers==8.12.43
pytz==2021.3
sqlparse==0.4.2
